# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: AI 철학자 상담 앱 API
  description: 철학자 AI와 대화하는 상담 서비스 API (MVP)
  version: 1.0.0
  contact:
    name: AI Counseling App Team
    email: support@aicounseling.app
servers:
  - url: http://localhost:8080/api
    description: 개발 서버
  - url: https://api.aicounseling.app/api
    description: 운영 서버
...
paths:
...
  /counselors:
    get:
      tags: [counselors]
      summary: 상담사 목록 조회
      parameters:
        - name: sort
          in: query
          schema:
            type: string
            enum: [popular, rating, recent]
            default: popular
      responses:
        200:
          description: 상담사 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CounselorListItem'
  /counselors/{id}:
    get:
      tags: [counselors]
      summary: 상담사 상세 정보
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 상담사 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CounselorDetail'
        404:
          description: 상담사를 찾을 수 없음
  /counselors/favorites:
    get:
      tags: [counselors]
      summary: 즐겨찾기 상담사 목록
      security:
        - bearerAuth: []
      responses:
        200:
          description: 즐겨찾기 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CounselorListItem'
  /counselors/{id}/favorite:
    post:
      tags: [counselors]
      summary: 상담사 즐겨찾기 추가
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        201:
          description: 즐겨찾기 추가됨
        409:
          description: 이미 즐겨찾기됨

    delete:
      tags: [counselors]
      summary: 상담사 즐겨찾기 제거
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        204:
          description: 즐겨찾기 제거됨

  # ===== 세션 =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CounselorListItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        description:
          type: string
        averageRating:
          type: number
          format: float
        totalSessions:
          type: integer
    CounselorDetail:
      allOf:
        - $ref: '#/components/schemas/CounselorListItem'
        - type: object
          properties:
            personalityMatrix:
              type: object
              description: 상담사 성격 특성
            specialties:
              type: array
              items:
                type: string
...